+++ android/src/main/kotlin/com/posthog/flutter/PostHogVersion.kt
@@ -1,3 +1,3 @@
 package com.posthog.flutter
 
-internal val postHogVersion = "4.10.5"
+internal val postHogVersion = "4.10.6"