+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Next
 
-- fix: check if image size is valid before sending snapshot ([#173](https://github.com/PostHog/posthog-flutter/pull/173))
+- fix: check if image size is valid before sending snapshot ([#174](https://github.com/PostHog/posthog-flutter/pull/174))
 
 ## 4.10.5