+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 4.10.6
+
 - fix: check if image size is valid before sending snapshot ([#174](https://github.com/PostHog/posthog-flutter/pull/174))
 
 ## 4.10.5